@@ -33,122 +34,121 @@
   },
   "homepage": "https://github.com/hiroppy/ssr-sample#readme",
   "devDependencies": {
-    "@babel/core": "7.5.4",
-    "@babel/plugin-proposal-class-properties": "7.5.0",
-    "@babel/plugin-proposal-object-rest-spread": "7.5.4",
+    "@apollo/react-testing": "^3.1.3",
+    "@babel/cli": "^7.7.0",
+    "@babel/core": "^7.7.2",
+    "@babel/node": "^7.7.0",
+    "@babel/plugin-proposal-class-properties": "^7.7.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
     "@babel/plugin-syntax-dynamic-import": "7.2.0",
-    "@babel/preset-env": "7.5.4",
-    "@babel/preset-react": "7.0.0",
-    "@babel/preset-typescript": "7.3.3",
-    "@storybook/addon-knobs": "5.1.9",
-    "@storybook/addon-storyshots": "5.1.9",
-    "@storybook/addon-viewport": "5.1.9",
-    "@storybook/cli": "5.1.9",
-    "@storybook/react": "5.1.9",
+    "@babel/preset-env": "^7.7.1",
+    "@babel/preset-react": "^7.7.0",
+    "@babel/preset-typescript": "^7.7.2",
+    "@loadable/babel-plugin": "^5.10.3",
+    "@loadable/webpack-plugin": "^5.7.1",
+    "@storybook/addon-knobs": "^5.2.6",
+    "@storybook/addon-storyshots": "^5.2.6",
+    "@storybook/addon-viewport": "^5.2.6",
+    "@storybook/cli": "^5.2.6",
+    "@storybook/react": "^5.2.6",
     "@storybook/storybook-deployer": "2.8.1",
-    "@types/body-parser": "1.17.0",
-    "@types/chance": "1.0.1",
-    "@types/dotenv": "6.1.1",
-    "@types/enzyme": "3.10.2",
-    "@types/enzyme-adapter-react-16": "1.0.5",
-    "@types/express": "4.17.0",
-    "@types/graphql": "14.2.2",
-    "@types/helmet": "0.0.43",
-    "@types/jest": "24.0.11",
+    "@testing-library/react": "^9.3.2",
+    "@testing-library/react-hooks": "^3.2.1",
+    "@types/body-parser": "^1.17.1",
+    "@types/compression": "^1.0.1",
+    "@types/dotenv": "^8.2.0",
+    "@types/express": "^4.17.2",
+    "@types/graphql": "^14.5.0",
+    "@types/helmet": "0.0.45",
+    "@types/jest": "^24.0.23",
+    "@types/loadable__component": "^5.10.0",
+    "@types/loadable__server": "^5.9.1",
     "@types/nanoid": "2.1.0",
-    "@types/nock": "10.0.3",
-    "@types/node": "10.14.4",
-    "@types/prop-types": "15.7.0",
-    "@types/react": "16.8.12",
-    "@types/react-dom": "16.8.3",
-    "@types/react-helmet": "5.0.8",
-    "@types/react-loadable": "5.5.1",
-    "@types/react-redux": "7.1.1",
-    "@types/react-router-dom": "4.3.1",
-    "@types/recompose": "0.30.5",
-    "@types/redux-mock-store": "1.0.0",
-    "@types/storybook__addon-storyshots": "4.0.0",
-    "@types/storybook__react": "4.0.1",
-    "@types/styled-components": "4.1.13",
+    "@types/nock": "^11.1.0",
+    "@types/node": "^12.12.11",
+    "@types/react": "^16.9.11",
+    "@types/react-dom": "^16.9.4",
+    "@types/react-helmet": "^5.0.14",
+    "@types/react-redux": "^7.1.5",
+    "@types/react-router-dom": "^5.1.2",
+    "@types/redux-mock-store": "^1.0.1",
+    "@types/storybook__addon-storyshots": "^5.1.1",
+    "@types/storybook__react": "^4.0.2",
+    "@types/styled-components": "^4.4.0",
+    "@types/supertest": "^2.0.8",
     "@types/ua-parser-js": "0.7.33",
-    "@types/webpack": "4.4.27",
-    "@types/webpack-env": "1.14.0",
-    "autocannon": "4.0.0",
-    "babel-jest": "24.8.0",
-    "babel-loader": "8.0.5",
-    "babel-plugin-macros": "2.6.1",
+    "@types/webpack-env": "^1.14.1",
+    "@typescript-eslint/eslint-plugin": "^2.8.0",
+    "@typescript-eslint/parser": "^2.8.0",
+    "autocannon": "^4.4.0",
+    "babel-jest": "^24.9.0",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-macros": "^2.7.0",
     "case-sensitive-paths-webpack-plugin": "2.2.0",
-    "chance": "1.0.18",
-    "clinic": "4.1.0",
-    "enzyme": "3.10.0",
-    "enzyme-adapter-react-16": "1.14.0",
-    "enzyme-to-json": "3.3.5",
-    "fork-ts-checker-webpack-plugin": "1.4.0",
-    "husky": "3.0.2",
-    "jest": "24.8.0",
-    "lint-staged": "9.4.1",
-    "nock": "10.0.6",
+    "clinic": "^5.0.0",
+    "cross-env": "^6.0.3",
+    "dotenv-webpack": "^1.7.0",
+    "eslint": "^6.6.0",
+    "eslint-config-prettier": "^6.7.0",
+    "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-react": "^7.16.0",
+    "fork-ts-checker-webpack-plugin": "^3.1.0",
+    "husky": "^3.1.0",
+    "jest": "^24.9.0",
+    "lint-staged": "^9.4.3",
+    "node-mocks-http": "^1.8.0",
+    "nodemon": "^2.0.0",
     "npm-run-all": "4.1.5",
-    "prettier": "1.18.2",
+    "prettier": "^1.19.1",
+    "react-test-renderer": "^16.12.0",
     "redux-devtools-extension": "2.13.8",
     "redux-mock-store": "1.5.3",
-    "redux-saga-test-plan": "3.7.0",
-    "require-context.macro": "1.0.4",
-    "rimraf": "2.6.3",
-    "ts-jest": "24.0.2",
-    "ts-node": "8.3.0",
-    "typescript": "3.5.3",
-    "webpack": "4.35.3",
-    "webpack-bundle-analyzer": "3.3.2",
-    "webpack-cli": "3.3.0",
-    "webpack-dev-middleware": "3.7.0",
-    "webpack-hot-middleware": "2.25.0",
-    "webpack-manifest-plugin": "2.0.4",
-    "webpack-merge": "4.2.1",
+    "redux-saga-test-plan": "^4.0.0-rc.3",
+    "require-context.macro": "^1.2.2",
+    "rimraf": "^3.0.0",
+    "robotstxt-webpack-plugin": "^7.0.0",
+    "supertest": "^4.0.2",
+    "typescript": "^3.7.2",
+    "webpack": "^4.41.2",
+    "webpack-bundle-analyzer": "^3.6.0",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-middleware": "^3.7.2",
+    "webpack-hot-middleware": "^2.25.0",
+    "webpack-manifest-plugin": "^2.2.0",
+    "webpack-merge": "^4.2.2",
+    "webpack-pwa-manifest": "^4.1.1",
     "workbox-webpack-plugin": "4.3.1"
   },
   "dependencies": {
-    "@octokit/rest": "16.28.4",
-    "apollo-boost": "0.4.3",
-    "apollo-link-schema": "1.2.2",
-    "apollo-server-express": "2.6.9",
+    "@loadable/component": "^5.10.3",
+    "@loadable/server": "^5.10.3",
+    "apollo-boost": "^0.4.4",
+    "apollo-link-schema": "^1.2.4",
+    "apollo-server-express": "^2.9.11",
     "body-parser": "1.19.0",
-    "connected-react-router": "4.5.0",
-    "core-js": "2.6.5",
-    "cross-env": "6.0.3",
-    "dotenv": "8.0.0",
-    "dotenv-webpack": "1.7.0",
+    "compression": "^1.7.4",
+    "core-js": "^3.4.1",
+    "cross-fetch": "^3.0.4",
+    "dotenv": "^8.2.0",
     "express": "4.17.1",
-    "graphql": "14.4.2",
-    "graphql-tools": "4.0.4",
-    "helmet": "3.18.0",
-    "history": "4.9.0",
-    "loadable-components": "2.2.3",
-    "nanoid": "2.0.1",
-    "react": "16.8.6",
-    "react-apollo": "2.5.4",
-    "react-dom": "16.8.6",
-    "react-helmet": "5.2.0",
-    "react-loadable": "5.5.0",
-    "react-loading": "2.0.3",
-    "react-redux": "5.1.1",
-    "react-router-dom": "4.3.1",
-    "recompose": "0.30.0",
-    "redux": "4.0.1",
-    "redux-saga": "0.16.2",
-    "styled-components": "4.3.2",
-    "ua-parser-js": "0.7.19",
-    "why-did-you-update": "1.0.6"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "lint-staged": {
-    "*.{js,ts,tsx,md,json,css}": [
-      "prettier --write",
-      "git add"
-    ]
+    "graphql": "^14.5.8",
+    "graphql-query-complexity": "^0.4.1",
+    "graphql-tools": "^4.0.6",
+    "helmet": "^3.21.2",
+    "history": "^4.10.1",
+    "nanoid": "^2.1.7",
+    "raw.macro": "^0.3.0",
+    "react": "^16.12.0",
+    "react-apollo": "^3.1.3",
+    "react-dom": "^16.12.0",
+    "react-helmet": "^5.2.1",
+    "react-markdown": "^4.2.2",
+    "react-redux": "^7.1.3",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
+    "redux": "^4.0.4",
+    "redux-saga": "^1.1.3",
+    "styled-components": "^4.4.1",
+    "ua-parser-js": "^0.7.20"
   }
 }
