@@ -1,5 +1,9 @@
 import * as actions from './pages';
 
+test('should create an action to set a base-url', () => {
+  expect(actions.setBaseUrl('url')).toMatchSnapshot();
+});
+
 test('should create an action to reset state', () => {
   expect(actions.resetPageStatus()).toMatchSnapshot();
 });
@@ -28,18 +32,26 @@ test('should create a failure action to load Top page', () => {
   expect(actions.loadTopPageFailure(new Error())).toMatchSnapshot();
 });
 
-test('should create an action to load Orgs page', () => {
-  expect(actions.loadOrgsPage('foo')).toMatchSnapshot();
+test('should create an action to load Saga page', () => {
+  expect(actions.loadSagaPage('2')).toMatchSnapshot();
 });
 
-test('should create a success action to load Orgs page', () => {
-  expect(actions.loadOrgsPageSuccess()).toMatchSnapshot();
+test('should create a success action to load Saga page', () => {
+  expect(actions.loadSagaPageSuccess()).toMatchSnapshot();
 });
 
-test('should create a failure action to load Orgs page', () => {
-  expect(actions.loadOrgsPageFailure(new Error())).toMatchSnapshot();
+test('should create a failure action to load Saga page', () => {
+  expect(actions.loadSagaPageFailure(new Error())).toMatchSnapshot();
 });
 
-test('should create an action to stop saga', () => {
-  expect(actions.stopSaga()).toMatchSnapshot();
+test('should create an action to load Apollo page', () => {
+  expect(actions.loadApolloPage()).toMatchSnapshot();
+});
+
+test('should create a success action to load Apollo page', () => {
+  expect(actions.loadApolloPageSuccess()).toMatchSnapshot();
+});
+
+test('should create a failure action to load Apollo page', () => {
+  expect(actions.loadApolloPageFailure(new Error())).toMatchSnapshot();
 });
