@@ -1,15 +1,14 @@
 import * as React from 'react';
 import Helmet from 'react-helmet';
 
-export interface Props {
+export type Props = {
   title: string;
-}
+  color?: string;
+};
 
-export const Head: React.SFC<Props> = ({ title }) => (
+export const Head: React.FC<Props> = ({ title, color = '#3498db' }) => (
   <Helmet>
-    <meta charSet="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>{title}</title>
-    <link href="https://fonts.googleapis.com/css?family=Muli" rel="stylesheet" />
+    <meta name="theme-color" content={color} />
   </Helmet>
 );
