@@ -1,11 +1,8 @@
 import * as React from 'react';
+import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-
-export interface Props {
-  userName: string;
-  orgName: string;
-}
+import { getEnv } from '../../selectors';
 
 const A = styled(Link)`
   color: #fff;
@@ -21,16 +18,20 @@ const HeaderComponent = styled.header`
   grid-template-columns: repeat(auto-fit, 200px);
   justify-content: space-between;
   padding: 15px 30px;
-
   & > * {
     text-align: center;
   }
 `;
 
-export const Header: React.SFC<Props> = ({ userName, orgName }) => (
-  <HeaderComponent>
-    <A to="/">{'<'}</A>
-    <span>{orgName}</span>
-    <span>{userName}</span>
-  </HeaderComponent>
-);
+export const Header: React.FC = () => {
+  const env = useSelector(getEnv);
+
+  return (
+    <HeaderComponent>
+      <A to="/">{'<'}</A>
+      <A to="/saga">saga</A>
+      <A to="/apollo">apollo</A>
+      <span>{env}</span>
+    </HeaderComponent>
+  );
+};
