@@ -1,20 +1,18 @@
-import * as React from 'react';
-import { storiesOf } from '@storybook/react';
-import { MemoryRouter } from 'react-router-dom';
-import { NotFound } from '.';
+import React from 'react';
+import { Provider } from 'react-redux';
+import configureStore from 'redux-mock-store';
+import NotFound from '.';
+import { initialState } from '../../../reducers';
 
-const stories = storiesOf('components/pages/NotFound', module);
+export default {
+  component: NotFound,
+  title: 'NotFound'
+};
 
-stories.add('default', () => (
-  <div
-    style={{
-      width: '80%',
-      margin: '50px auto',
-      background: '#f5f5f5'
-    }}
-  >
-    <MemoryRouter initialEntries={['/']} keyLength={0}>
-      <NotFound />
-    </MemoryRouter>
-  </div>
-));
+const store = configureStore()(initialState);
+
+export const Base = () => (
+  <Provider store={store}>
+    <NotFound />
+  </Provider>
+);
