@@ -1,11 +1,20 @@
-import * as React from 'react';
-import { shallow } from 'enzyme';
-import { NotFound } from '.';
+import React from 'react';
+import { render } from '@testing-library/react';
+import { Provider } from 'react-redux';
+import configureStore from 'redux-mock-store';
+import NotFound from '.';
+import { initialState } from '../../../reducers';
+import * as actions from '../../../actions/pages';
 
-test('should call stopSaga', () => {
-  const stop = jest.fn();
+test('should call loadTopPage via useDispatch', () => {
+  const loadTopPage = jest.spyOn(actions, 'loadTopPage');
+  const mockStore = configureStore()(initialState);
 
-  shallow(<NotFound stopSaga={stop} />);
+  render(
+    <Provider store={mockStore}>
+      <NotFound />
+    </Provider>
+  );
 
-  expect(stop).toHaveBeenCalled();
+  expect(loadTopPage).toBeCalled();
 });
