@@ -1,1 +1,3 @@
-export * from './NotFound';
+import { NotFound } from './NotFound';
+
+export default NotFound;
