@@ -1,18 +1,26 @@
 import { combineReducers } from 'redux';
-import { RouterState } from 'connected-react-router';
-import { reducer as usersReducer, State as UsersState } from './users';
-import { reducer as orgsReducer, State as OrgsState } from './orgs';
-import { reducer as pagesReducer, State as PagesState } from './pages';
+import {
+  reducer as pagesReducer,
+  initialState as pagesInitialState,
+  State as PagesState
+} from './pages';
+import {
+  reducer as sagaPageReducer,
+  initialState as sagaPageInitialState,
+  State as SagaPageState
+} from './sagaPage';
 
-export interface State {
-  users: UsersState;
-  orgs: OrgsState;
+export type State = {
   pages: PagesState;
-  router: RouterState;
-}
+  sagaPage: SagaPageState;
+};
+
+export const initialState = {
+  pages: pagesInitialState,
+  sagaPage: sagaPageInitialState
+};
 
 export const rootReducer = combineReducers({
-  users: usersReducer,
-  orgs: orgsReducer,
-  pages: pagesReducer
+  pages: pagesReducer,
+  sagaPage: sagaPageReducer
 });
