@@ -1,5 +1,5 @@
 import { rootSaga } from '.';
 
 test('should register sagas', () => {
-  expect(rootSaga().next().value.ALL).toHaveLength(2);
+  expect((rootSaga().next().value as any).payload).toHaveLength(2);
 });
