@@ -1,10 +1,10 @@
 import { all, fork } from 'redux-saga/effects';
 import { pagesProcess } from './pages';
-import { orgsProcess } from './orgs';
+import { fetchSagaProcess } from './fetchSaga';
 
 /**
  * Root for saga
  */
 export function* rootSaga() {
-  yield all([fork(pagesProcess), fork(orgsProcess)]);
+  yield all([fork(pagesProcess), fork(fetchSagaProcess)]);
 }
