@@ -1,42 +1,18 @@
 import { END } from 'redux-saga';
 import { expectSaga } from 'redux-saga-test-plan';
 import { pagesProcess } from './pages';
-import {
-  loadAppProcess,
-  loadAppProcessSuccess,
-  loadAppProcessFailure,
-  loadOrgsPage,
-  loadOrgsPageSuccess,
-  loadOrgsPageFailure,
-  loadTopPage,
-  loadTopPageSuccess,
-  stopSaga,
-  resetPageStatus
-} from '../actions/pages';
-import { setUserName } from '../actions/users';
-import { resetOrgs, fetchRepos, fetchReposSuccess, fetchReposFailure } from '../actions/orgs';
-
-afterEach(() => {
-  delete process.env.IS_BROWSER;
-});
-
-const initialState = {
-  isLoadingCompletion: false,
-  orgs: {
-    name: 'foo'
-  },
-  users: {
-    name: ''
-  }
-};
+import * as actions from '../actions/pages';
+import * as fetchSagaActions from '../actions/fetchSaga';
+import { initialState } from '../reducers/pages';
+import { sagaSamples } from '../../server/responseSchema';
 
 describe('LOAD_APP_PROCESS', () => {
   test('should call LOAD_APP_PROCESS_SUCCESS action', () => {
     return expectSaga(pagesProcess)
       .withState(initialState)
-      .put(setUserName('hiroppy'))
-      .put(loadAppProcessSuccess())
-      .dispatch(loadAppProcess())
+      .put(actions.setEnv('test'))
+      .put(actions.loadAppProcessSuccess())
+      .dispatch(actions.loadAppProcess())
       .silentRun();
   });
 });
@@ -45,9 +21,9 @@ describe('LOAD_TOP_PAGE', () => {
   test('should call LOAD_TOP_PAGE_SUCCESS with SSR', () => {
     return expectSaga(pagesProcess)
       .withState(initialState)
-      .put(loadTopPageSuccess())
+      .put(actions.loadTopPageSuccess())
       .put(END)
-      .dispatch(loadTopPage())
+      .dispatch(actions.loadTopPage())
       .silentRun();
   });
 
@@ -56,79 +32,62 @@ describe('LOAD_TOP_PAGE', () => {
 
     return expectSaga(pagesProcess)
       .withState(initialState)
-      .put(loadTopPageSuccess())
-      .dispatch(loadTopPage())
+      .put(actions.loadTopPageSuccess())
+      .dispatch(actions.loadTopPage())
       .silentRun();
   });
 });
 
-describe('LOAD_ORGS_PAGE', () => {
-  test('should call LOAD_ORGS_PAGE_SUCCESS action with SSR', () => {
+describe('LOAD_SAGA_PAGE', () => {
+  test('should call LOAD_SAGA_PAGE_SUCCESS with SSR', () => {
     return expectSaga(pagesProcess)
       .withState(initialState)
-      .put(loadOrgsPageSuccess())
+      .put(actions.loadSagaPageSuccess())
       .put(END)
-      .dispatch(loadOrgsPage('foo')) // make it the same name as initialState
+      .dispatch(actions.loadSagaPage('1'))
+      .dispatch(fetchSagaActions.fetchSagaCodeSuccess(sagaSamples))
       .silentRun();
   });
 
-  test('should call LOAD_ORGS_PAGE_SUCCESS action with CSR', () => {
-    process.env.IS_BROWSER = 'true';
-
+  test('should call LOAD_SAGA_PAGE_FAILURE with SSR', () => {
     return expectSaga(pagesProcess)
       .withState(initialState)
-      .put(resetOrgs())
-      .put(fetchRepos('bar'))
-      .put(loadOrgsPageSuccess())
-      .dispatch(loadOrgsPage('bar'))
-      .dispatch(
-        fetchReposSuccess({
-          name: 'bar',
-          repos: []
-        })
-      )
-      .silentRun();
-  });
-
-  test('should call LOAD_ORGS_PAGE_FAILURE action when getting 404 with CSR', () => {
-    process.env.IS_BROWSER = 'true';
-
-    return expectSaga(pagesProcess)
-      .withState(initialState)
-      .put(resetOrgs())
-      .put(fetchRepos('bar'))
-      .dispatch(loadOrgsPage('bar'))
-      .dispatch(fetchReposFailure(new Error('404')))
-      .silentRun();
-  });
-
-  test('should call LOAD_ORGS_PAGE action when getting 403 with CSR', () => {
-    process.env.IS_BROWSER = 'true';
-
-    return expectSaga(pagesProcess)
-      .withState(initialState)
-      .put(resetOrgs())
-      .put(fetchRepos('bar'))
-      .dispatch(loadOrgsPage('bar'))
-      .dispatch(fetchReposFailure(new Error('403')))
-      .silentRun();
-  });
-});
-
-describe('STOP_SAGA', () => {
-  test('should call END action', () => {
-    return expectSaga(pagesProcess)
+      .put(actions.loadSagaPageFailure(new Error()))
       .put(END)
-      .dispatch(stopSaga())
+      .dispatch(actions.loadSagaPage('1'))
+      .dispatch(fetchSagaActions.fetchSagaCodeFailure(new Error()))
+      .silentRun();
+  });
+
+  test("shouldn't call END when CSR", () => {
+    process.env.IS_BROWSER = 'true';
+
+    return expectSaga(pagesProcess)
+      .withState(initialState)
+      .put(actions.loadSagaPageSuccess())
+      .dispatch(actions.loadSagaPage('1'))
+      .dispatch(fetchSagaActions.fetchSagaCodeSuccess(sagaSamples))
       .silentRun();
   });
 });
 
-describe('@@router/LOCATION_CHANGE', () => {
-  test('should call RESET_PAGE_STATUS action', () => {
+describe('LOAD_APOLLO_PAGE', () => {
+  test('should call LOAD_APOLLO_PAGE_SUCCESS with SSR', () => {
     return expectSaga(pagesProcess)
-      .put(resetPageStatus())
-      .dispatch({ type: '@@router/LOCATION_CHANGE' })
+      .withState(initialState)
+      .put(actions.loadApolloPageSuccess())
+      .put(END)
+      .dispatch(actions.loadApolloPage())
+      .silentRun();
+  });
+
+  test('should call LOAD_APOLLO_PAGE_SUCCESS action with CSR', () => {
+    process.env.IS_BROWSER = 'true';
+
+    return expectSaga(pagesProcess)
+      .withState(initialState)
+      .put(actions.loadApolloPageSuccess())
+      .dispatch(actions.loadApolloPage())
       .silentRun();
   });
 });
