@@ -1,40 +1,44 @@
-import { createStore, applyMiddleware } from 'redux';
-import createSagaMiddleware from 'redux-saga';
-import { connectRouter, routerMiddleware } from 'connected-react-router';
+import { createStore, applyMiddleware, Store } from 'redux';
+import createSagaMiddleware, { SagaMiddleware } from 'redux-saga';
 import { createBrowserHistory, createMemoryHistory } from 'history';
 import { composeWithDevTools } from 'redux-devtools-extension/developmentOnly';
 import { rootReducer } from '../reducers';
 import { rootSaga } from '../sagas';
 
 export const history = process.env.IS_BROWSER ? createBrowserHistory() : createMemoryHistory();
-const sagaMiddleware = createSagaMiddleware();
 
-const createEnhancer = () => {
+const createEnhancer = (sagaMiddleware: ReturnType<typeof createSagaMiddleware>) => {
   const composeEnhancers = composeWithDevTools({});
 
-  return composeEnhancers(applyMiddleware(sagaMiddleware, routerMiddleware(history)));
+  return composeEnhancers(applyMiddleware(sagaMiddleware));
 };
 
-export const runSaga = async () => {
-  return sagaMiddleware.run(rootSaga).done;
-};
+export const configureStore = (
+  preloadedState: Record<string, any> = {}
+): {
+  store: Store;
+  runSaga: () => Promise<SagaMiddleware<typeof rootSaga>['run']>;
+} => {
+  const sagaMiddleware = createSagaMiddleware();
+  const enhancer = createEnhancer(sagaMiddleware);
+  const store = createStore(rootReducer, preloadedState, enhancer);
+  const runSaga = async () => {
+    return sagaMiddleware.run(rootSaga).toPromise();
+  };
 
-export const configureStore = (preloadedState: Object = {}) => {
-  const enhancer = createEnhancer();
-  const store = createStore(connectRouter(history)(rootReducer), preloadedState, enhancer);
-
-  runSaga();
+  // for client-side
+  if (process.env.IS_BROWSER) {
+    runSaga();
+  }
 
   /* istanbul ignore next */
   if (module.hot) {
-    module.hot.accept('../reducers', () => {
-      const {
-        rootReducer: nextReducer
-      }: { rootReducer: typeof rootReducer } = require('../reducers');
+    module.hot.accept('../reducers', async () => {
+      const { rootReducer } = await import(/* webpackMode: "eager" */ '../reducers');
 
-      store.replaceReducer(connectRouter(history)(nextReducer));
+      store.replaceReducer(rootReducer);
     });
   }
 
-  return store;
+  return { store, runSaga };
 };
