@@ -1,90 +1,41 @@
 import { makeExecutableSchema, IResolvers } from 'graphql-tools';
-import Octokit from '@octokit/rest';
-
-const octokit = new Octokit();
-
-export type Organizations = Array<{
-  name: string;
-  uri: string;
-  uid: number;
-}>;
-
-export const organizations: Organizations = [
-  {
-    name: 'nodejs',
-    uri: 'https://github.com/nodejs',
-    uid: 1
-  },
-  {
-    name: 'facebook',
-    uri: 'https://github.com/facebook',
-    uid: 2
-  },
-  {
-    name: 'google',
-    uri: 'https://github.com/google',
-    uid: 3
-  },
-  {
-    name: 'microsoft',
-    uri: 'https://github.com/microsoft',
-    uid: 4
-  }
-];
-
-export type Author = {
-  id: number;
-  name: string;
-  avatar_url: string;
-  html_url: string;
-  blog: string;
-};
+import { apolloSamples } from '../server/responseSchema';
 
 export const typeDefs = `
-  type Organization {
-    name: String!
-    uri: String!
-    uid: Int!
-  }
-
-  type Author {
+  type Sample {
     id: Int!
+    code: String!
     name: String!
-    avatar_url: String!
-    html_url: String
-    blog: String!
+    likeCount: Int!
+    description: String
   }
 
   type Query {
-    organizations: [Organization]
-    organization(name: String!): Organization
-    author: Author
+    samples(maxLength: Int): [Sample]
   }
 
   type Mutation {
-    addOrganization(name: String!): Organization
+    addLike(id: Int): Sample
   }
 `;
 
 export const resolvers: IResolvers = {
   Query: {
-    organizations: () => organizations,
-    organization: (obj, { name }) => organizations.find((o) => o.name === name),
-    author: () => octokit.users.getByUsername({ username: 'hiroppy' }).then(({ data }) => data)
+    samples: (obj, { maxLength }: { maxLength?: number }) => {
+      return apolloSamples.slice(0, maxLength || apolloSamples.length);
+    }
   },
   Mutation: {
-    addOrganization: (obj, { name }) => {
-      if (organizations.some((o) => o.name === name)) throw new Error('already exist');
+    addLike: (obj, { id }: { id: number }) => {
+      const item = apolloSamples.find(({ id: fileId }) => id === fileId);
 
-      const org = {
-        name,
-        uid: organizations.length + 1,
-        uri: `https://github.com/${name}`
-      };
+      if (item) {
+        ++item.likeCount;
 
-      organizations.push(org);
-
-      return org;
+        return item;
+      } else {
+        throw new Error('Could not find');
+      }
     }
   }
 };
