@@ -1,57 +1,43 @@
-// renderToNodeStream
-// https://gist.github.com/hiroppy/1c89d73a12073bad0c187aaab4ca92c2
-
+import { resolve } from 'path';
 import { Request, Response } from 'express';
-import * as React from 'react';
+import React from 'react';
 import { Provider } from 'react-redux';
-import { renderToString } from 'react-dom/server';
+import { renderToString, renderToStaticMarkup } from 'react-dom/server';
 import { StaticRouter } from 'react-router-dom';
 import Helmet from 'react-helmet';
 import { ServerStyleSheet } from 'styled-components';
-//@ts-ignore
-import { getLoadableState } from 'loadable-components/server';
-
-// graphql
-import { ApolloProvider, getDataFromTree } from 'react-apollo';
-import { createClient } from '../../../graphql/client';
-
+import { ChunkExtractor } from '@loadable/server';
+import { ApolloProvider, getMarkupFromTree } from 'react-apollo';
+import { SchemaLink } from 'apollo-link-schema';
 import { renderFullPage } from '../../renderFullPage';
-import { Router } from '../../../client/Router';
-import { configureStore, runSaga } from '../../../client/store/configureStore';
+import { Router } from '../../../client/router/index';
+import { configureStore } from '../../../client/store/configureStore';
+import { createClient } from '../../../graphql/client';
+import { schema } from '../../../graphql/schema';
+import { setBaseUrl } from '../../../client/actions/pages';
 
-const fileName = 'main';
-const main = new RegExp(`^${fileName}~.*\.js$`);
-const vendor = new RegExp(`^vendor~${fileName}~.*\.js$`);
-
-// You need to reboot this server if you change client javascript files.
-// You need to read the manifest in `get` method if you do not want to restart.
-const assets =
-  process.env.NODE_ENV === 'production'
-    ? (() => {
-        const manifest: {
-          [key: string]: string;
-        } = require('../../../../client/manifest');
-
-        const entryPoints = [manifest['manifest.js']];
-
-        for (const [key, value] of Object.entries(manifest)) {
-          if (main.test(key) || vendor.test(key)) entryPoints.push(value);
-        }
-
-        return entryPoints;
-      })()
-    : [`/public/${fileName}.bundle.js`];
+const statsFile = resolve(
+  __dirname,
+  process.env.NODE_ENV !== 'production'
+    ? '../../../../dist/client/loadable-stats.json'
+    : '../../../../client/loadable-stats.json'
+);
 
 export async function get(req: Request, res: Response) {
+  const baseUrl = `${req.protocol}://${req.get('Host')}`;
   const { nonce }: { nonce: string } = res.locals;
-  const store = configureStore();
-  const client = createClient();
+  const { store, runSaga } = configureStore();
+  const client = createClient({ link: new SchemaLink({ schema }) });
   const sheet = new ServerStyleSheet();
+  const context = {};
 
-  const App = (
+  // for Node.js because `fetch` requires absolute URLs
+  store.dispatch(setBaseUrl(baseUrl));
+
+  const App = () => (
     <ApolloProvider client={client}>
       <Provider store={store}>
-        <StaticRouter location={req.url} context={{}}>
+        <StaticRouter location={req.url} context={context}>
           {/* add `div` because of `hydrate` */}
           <div id="root">
             <Router />
@@ -62,28 +48,32 @@ export async function get(req: Request, res: Response) {
   );
 
   try {
-    const [loadableState] = await Promise.all([
-      getLoadableState(App), // kick redux-saga and styled-components
-      runSaga(),
-      getDataFromTree(App)
+    const extractor = new ChunkExtractor({ statsFile });
+    const tree = extractor.collectChunks(<App />);
+
+    await Promise.all([
+      // kick apollo, redux-saga, and styled-components
+      getMarkupFromTree({
+        tree,
+        renderFunction: renderToStaticMarkup
+      }),
+      runSaga()
     ]);
 
+    const body = renderToString(tree);
     const preloadedState = JSON.stringify(store.getState());
     const helmetContent = Helmet.renderStatic();
     const meta = `
       ${helmetContent.meta.toString()}
       ${helmetContent.title.toString()}
-      ${helmetContent.link.toString()}
     `.trim();
     const style = sheet.getStyleTags();
-    const body = renderToString(App);
-    const scripts = loadableState.getScriptTag().replace(/^\<script\>/, `<script nonce=${nonce}>`);
+    const scripts = extractor.getScriptTags({ nonce });
     const graphql = JSON.stringify(client.extract());
 
-    res.send(
-      renderFullPage({ meta, assets, body, style, preloadedState, scripts, graphql, nonce })
-    );
+    return res.send(renderFullPage({ meta, body, style, preloadedState, scripts, graphql, nonce }));
   } catch (e) {
-    res.status(500).send(e.message);
+    console.error(e);
+    return res.status(500).send(e.message);
   }
 }
