@@ -3,7 +3,6 @@ import { renderFullPage } from './renderFullPage';
 test('should create safe html', () => {
   const meta = '<meta>meta</meta>';
   const style = '<style>style</style>';
-  const assets = ['script'];
   const body = '<body>body</body>';
   const preloadedState = JSON.stringify({
     arr: ['foo', 'bar'],
@@ -18,6 +17,6 @@ test('should create safe html', () => {
   const nonce = 'nonce';
 
   expect(
-    renderFullPage({ meta, style, assets, body, preloadedState, scripts, graphql, nonce })
+    renderFullPage({ meta, style, body, preloadedState, scripts, graphql, nonce })
   ).toMatchSnapshot();
 });
